# api/openapi.yaml
openapi: 3.0.3
info:
  title: guided-ssh API
  description: |
    REST API of the guided-ssh server: certificate issuance (human + CI),
    host enrollment, grant management, and the read/audit endpoints of the
    web UI. This specification is the single source of truth; the
    Angular API client is generated from it (`make web-api`).

    Authentication: OIDC ID token as bearer token. Roles of the admin API
    (phase 8): `admin` (everything), `auditor` (read/export audit + read),
    `readonly` (read resources); roles are mapped to IdP groups and are
    inclusive of one another (admin ⊃ auditor ⊃ readonly).
  version: "1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
...
  /v1/ca/bundle/{purpose}:
    get:
      tags: [ca]
      summary: CA bundle in authorized_keys format
      security: []
      operationId: getCaBundle
      parameters:
        - name: purpose
          in: path
          required: true
          schema: { type: string, enum: [user, host] }
      responses:
        "200":
          description: CA bundle
          content:
            text/plain:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sign/user:
    post:
      tags: [sign]
      summary: Exchange ID token for SSH user certificate
      operationId: signUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignRequest" }
      responses:
        "200":
          description: Signed certificate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/sign/ci:
    post:
      tags: [sign]
      summary: Exchange GitLab job token for CI certificate
      operationId: signCi
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignRequest" }
      responses:
        "200":
          description: Signed certificate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/enroll:
    post:
      tags: [enrollment]
      summary: Host enrollment against a one-time token
      security: []
      operationId: enrollHost
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnrollRequest" }
      responses:
        "200":
          description: Host certificate and mTLS material
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnrollResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OIDC-ID-Token
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain: { schema: { type: string } }
    Unauthorized:
      description: Token missing or invalid
      content:
        text/plain: { schema: { type: string } }
    Forbidden:
      description: Insufficient role/permission
      content:
        text/plain: { schema: { type: string } }
    NotFound:
      description: Not found
      content:
        text/plain: { schema: { type: string } }
...
  schemas:
...
    SignRequest:
      type: object
      required: [public_key]
      properties:
        public_key:
          type: string
          description: SSH public key in authorized_keys format
        validity_seconds:
          type: integer
          format: int64
          description: Requested validity; capped by policy and grants
    SignResponse:
      type: object
      required: [certificate, serial, key_id, principals, valid_after, valid_before]
      properties:
        certificate: { type: string }
        serial: { type: integer, format: int64 }
        key_id: { type: string }
        principals: { type: array, items: { type: string } }
        valid_after: { type: string, format: date-time }
        valid_before: { type: string, format: date-time }
    EnrollRequest:
      type: object
      required: [token, hostname, ssh_public_key, mtls_csr]
      properties:
        token: { type: string }
        hostname: { type: string }
        ssh_public_key: { type: string }
        mtls_csr: { type: string, description: PEM-encoded CSR for the mTLS client certificate }
        tags:
          type: object
          additionalProperties: { type: string }
    EnrollResponse:
      type: object
      required: [host_id, host_certificate, valid_before, user_ca_bundle, mtls_certificate, mtls_ca]
      properties:
        host_id: { type: string, format: uuid }
        host_certificate: { type: string }
        valid_before: { type: string, format: date-time }
        user_ca_bundle: { type: string }
        mtls_certificate: { type: string }
        mtls_ca: { type: string }
...